Minor updates that have collected as things have gone along.
Updated doctags to include perforce. Added some snippets for testing using
Google Test and for creating Doxygen comments. Tweaked the Cpp and Header vim
commands to allow opening the file in the current window instead of opening a
new one. Added a command to source .vimrc if I make changes.